+++ src/main/java/de/samply/spring/MtbaConst.java
@@ -10,6 +10,7 @@ public class MtbaConst {
   public final static String ID_MANAGER_API_KEY = "ID_MANAGER_API_KEY";
   public final static String ID_MANAGER_URL = "ID_MANAGER_URL";
   public final static String ID_MANAGER_PSEUDONYM_ID_TYPE = "ID_MANAGER_PSEUDONYM_ID_TYPE";
+  public final static String ID_MANAGER_PAGE_SIZE = "ID_MANAGER_PAGE_SIZE";
   public final static String LOCAL_ID_CSV_FILENAME = "LOCAL_ID_CSV_FILENAME";
   public final static String LOCAL_ID_CSV_LOCAL_ID_HEADER = "LOCAL_ID_CSV_LOCAL_ID_HEADER";
   public final static String IDAT_CSV_FILENAME = "IDAT_CSV_FILENAME";
@@ -45,6 +46,7 @@ public class MtbaConst {
   public final static String IDAT_CSV_GENDER_HEADER_SV = HEAD_SV + IDAT_CSV_GENDER_HEADER + ":#{null}" + BOTTOM_SV;
   public final static String FHIR_BUNDLE_CSV_FILENAME_SV = HEAD_SV + FHIR_BUNDLE_CSV_FILENAME + ":#{'data_mutations_extended.txt'}" + BOTTOM_SV;
 
+  public final static String ID_MANAGER_PAGE_SIZE_SV = HEAD_SV + ID_MANAGER_PAGE_SIZE + ":#{'10'}" + BOTTOM_SV;
 
 
   // Camunda BP Constants